# server/openapi/swagger.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Traveling Cook Calculator API
  description: |
    API for managing Cook and Run events - a social cooking experience where teams 
    cook different courses at different locations and move between them.

    ## Authentication
    All endpoints require a valid JWT Bearer token. The JWT should contain:
    - `sub`: User ID
    - `role`: User role (admin, organizer, participant)
    - `exp`: Token expiration time

    ## Rate Limiting
    - 1000 requests per hour per authenticated user
    - 100 requests per hour for unauthenticated endpoints

    ## Error Handling
    The API uses conventional HTTP response codes and returns error details in JSON format.
    All error responses follow the same structure with error codes, messages, and optional details.
servers:
  - url: http://0.0.0.0:3000/
    description: Local development server
security:
  - bearerAuth: []
paths:
  /cook_and_run:
    get:
      summary: List all cook and run projects
      description: |
        Returns a paginated list of cook and run projects for the authenticated user.
        Results are sorted by creation date (newest first) by default.
      parameters:
        - name: userId
          in: query
          required: true
          description: ID of the user whose projects to retrieve
          schema:
            type: string
            maxLength: 64
            pattern: "^[a-zA-Z0-9_-]+$"
          example: "user123"
        - name: page
          in: query
          required: false
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
        - name: limit
          in: query
          required: false
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          example: 20
        - name: sort
          in: query
          required: false
          description: Sort field and direction
          schema:
            type: string
            enum:
              [
                created_asc,
                created_desc,
                name_asc,
                name_desc,
                edited_asc,
                edited_desc,
              ]
            default: created_desc
      responses:
        "200":
          $ref: "#/components/responses/CookAndRunMetaListResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "429":
          $ref: "#/components/responses/TooManyRequestsError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Cook and Run
  /cook_and_run/{cookAndRunId}:
    post:
      summary: Create a new cook and run project
      description: Creates a new cook and run project with the specified ID
      parameters:
        - $ref: "#/components/parameters/CookAndRunId"
      requestBody:
        $ref: "#/components/requestBodies/CookAndRunCreateRequest"
      responses:
        "201":
          description: Cook and Run created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CookAndRun"
              examples:
                new_project:
                  summary: Newly created project
                  value:
                    id: "123e4567-e89b-12d3-a456-426614174000"
                    name: "Summer Cook & Run 2024"
                    created: "2024-03-15T10:00:00Z"
                    edited: "2024-03-15T10:00:00Z"
                    occur: "2024-06-15T18:00:00Z"
                    team_list: []
                    course_list: []
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "409":
          description: Cook and Run with this ID already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "RESOURCE_CONFLICT"
                message: "Cook and Run project with this ID already exists"
                timestamp: "2024-03-15T10:00:00Z"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Cook and Run

    get:
      summary: Get cook and run project details
      description: Returns detailed information about a specific cook and run project
      parameters:
        - $ref: "#/components/parameters/CookAndRunId"
      responses:
        "200":
          $ref: "#/components/responses/CookAndRunResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Cook and Run

    delete:
      summary: Delete cook and run project
      description: |
        Permanently deletes a cook and run project and all associated data.
        This action cannot be undone.
      parameters:
        - $ref: "#/components/parameters/CookAndRunId"
      responses:
        "204":
          description: Cook and Run deleted successfully
        "400":
          description: Cannot delete project with registered teams (use force=true to override)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Cook and Run
  /cook_and_run/{cookAndRunId}/name:
    patch:
      summary: Update cook and run project name
      description: Updates only the name of a cook and run project
      parameters:
        - $ref: "#/components/parameters/CookAndRunId"
      requestBody:
        $ref: "#/components/requestBodies/CookAndRunNameUpdateRequest"
      responses:
        "200":
          description: Cook and Run name updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  name:
                    type: string
                  edited:
                    type: string
                    format: date-time
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Cook and Run
  /cook_and_run/{cookAndRunId}/start_point:
    patch:
      summary: Update start point
      description: Updates the start point location for cook and run project
      parameters:
        - $ref: "#/components/parameters/CookAndRunId"
      requestBody:
        $ref: "#/components/requestBodies/CookAndRunLocationUpdateRequest"
      responses:
        "200":
          description: Start point updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Locations
    delete:
      summary: Delete start point
      description: Removes the start point location from the cook and run project
      parameters:
        - $ref: "#/components/parameters/CookAndRunId"
      responses:
        "204":
          description: Start point deleted successfully
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Locations
  /cook_and_run/{cookAndRunId}/end_point:
    patch:
      summary: Update end point
      description: Updates the end point location for cook and run project
      parameters:
        - $ref: "#/components/parameters/CookAndRunId"
      requestBody:
        $ref: "#/components/requestBodies/CookAndRunLocationUpdateRequest"
      responses:
        "200":
          description: End point updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Locations
    delete:
      summary: Delete end point
      description: Removes the end point location from the cook and run project
      parameters:
        - $ref: "#/components/parameters/CookAndRunId"
      responses:
        "204":
          description: End point deleted successfully
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Locations
...
components:
  parameters:
    CookAndRunId:
      name: cookAndRunId
      in: path
      required: true
      description: Unique identifier for the cook and run project
      schema:
        type: string
        format: uuid
      example: "123e4567-e89b-12d3-a456-426614174000"
...
  requestBodies:
    CookAndRunCreateRequest:
      description: Request body to create a Cook and Run project
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CookAndRunCreateData"
          examples:
            basic_event:
              summary: Basic event creation
              value:
                name: "Neighborhood Cook & Run"
                userId: "user123"
            detailed_event:
              summary: Event with full details
              value:
                name: "Summer Cook & Run 2024"
                userId: "user123"
                occur: "2024-06-15T18:00:00Z"
                description: "Annual summer cooking event for the neighborhood"
...
    CookAndRunNameUpdateRequest:
      description: Update cook and run project name
      required: true
      content:
        application/json:
          schema:
            type: object
            required:
              - name
            properties:
              name:
                type: string
                minLength: 1
                maxLength: 64
                description: New name for the cook and run project
    CookAndRunLocationUpdateRequest:
      description: Update cook and run start or end point
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Address"
          examples:
            central_location:
              summary: Central city location
              value:
                address: "Marienplatz 1, 80331 München, Germany"
                latitude: 48.137154
                longitude: 11.575490
...
  responses:
    BadRequestError:
      description: Bad request - invalid input parameters or request format
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            invalid_uuid:
              summary: Invalid UUID format
              value:
                error: "BAD_REQUEST"
                message: "Invalid UUID format in path parameter"
                timestamp: "2024-03-15T10:00:00Z"
            missing_field:
              summary: Missing required field
              value:
                error: "BAD_REQUEST"
                message: "Missing required fields in request body"
                details:
                  - field: "name"
                    message: "Name is required and cannot be empty"
                timestamp: "2024-03-15T10:00:00Z"
    UnauthorizedError:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            missing_token:
              summary: Missing authentication token
              value:
                error: "UNAUTHORIZED"
                message: "Authentication token is required"
                timestamp: "2024-03-15T10:00:00Z"
            invalid_token:
              summary: Invalid or expired token
              value:
                error: "UNAUTHORIZED"
                message: "Authentication token is invalid or expired"
                timestamp: "2024-03-15T10:00:00Z"
    ForbiddenError:
      description: Access denied - insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            insufficient_permissions:
              summary: User lacks required permissions
              value:
                error: "FORBIDDEN"
                message: "You do not have permission to access this resource"
                timestamp: "2024-03-15T10:00:00Z"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            project_not_found:
              summary: Cook and Run project not found
              value:
                error: "NOT_FOUND"
                message: "Cook and Run project not found"
                timestamp: "2024-03-15T10:00:00Z"
            team_not_found:
              summary: Team not found
              value:
                error: "NOT_FOUND"
                message: "Team not found in this project"
                timestamp: "2024-03-15T10:00:00Z"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            validation_failed:
              summary: Multiple validation errors
              value:
                error: "VALIDATION_ERROR"
                message: "Request validation failed"
                details:
                  - field: "email"
                    message: "Invalid email format"
                  - field: "phone"
                    message: "Phone number must be in international format"
                  - field: "members"
                    message: "Number of members must be between 1 and 20"
                timestamp: "2024-03-15T10:00:00Z"
    TooManyRequestsError:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            rate_limit_exceeded:
              summary: Too many requests
              value:
                error: "TOO_MANY_REQUESTS"
                message: "Rate limit exceeded. Try again later."
                details:
                  - field: "retry_after"
                    message: "3600 seconds"
                timestamp: "2024-03-15T10:00:00Z"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            server_error:
              summary: Unexpected server error
              value:
                error: "INTERNAL_SERVER_ERROR"
                message: "An unexpected error occurred. Please try again later."
                timestamp: "2024-03-15T10:00:00Z"
    CookAndRunMetaListResponse:
      description: Paginated list of Cook and Run projects
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: array
                items:
                  $ref: "#/components/schemas/CookAndRunMeta"
              pagination:
                $ref: "#/components/schemas/PaginationInfo"
            required:
              - data
              - pagination
          examples:
            project_list:
              summary: List of projects with pagination
              value:
                data:
                  - id: "123e4567-e89b-12d3-a456-426614174000"
                    name: "Summer Cook & Run 2024"
                    created: "2024-03-01T10:00:00Z"
                    edited: "2024-03-15T14:30:00Z"
                  - id: "456e7890-e12b-34c5-d678-901234567890"
                    name: "Neighborhood Cooking Adventure"
                    created: "2024-02-15T09:00:00Z"
                    edited: "2024-03-10T16:45:00Z"
                pagination:
                  page: 1
                  limit: 20
                  total: 2
                  total_pages: 1
    CookAndRunResponse:
      description: Complete Cook and Run project details
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CookAndRun"
          examples:
            complete_project:
              summary: Complete project with teams and courses
              value:
                id: "123e4567-e89b-12d3-a456-426614174000"
                name: "Summer Cook & Run 2024"
                created: "2024-03-01T10:00:00Z"
                edited: "2024-03-15T14:30:00Z"
                occur: "2024-06-15T18:00:00Z"
                team_list:
                  - id: "789e0123-e45f-67g8-h901-234567890123"
                    name: "The Cooking Ninjas"
                    created_by_user: "user456"
                    created: "2024-03-10T12:00:00Z"
                    edited: "2024-03-10T12:00:00Z"
                    address:
                      address: "Maximilianstrasse 10, 80539 München, Germany"
                      latitude: 48.142385
                      longitude: 11.579381
                    needs_check: true
                    status: "approved"
                    note_list: []
                course_list:
                  - id: "456e7890-e12b-34c5-d678-901234567890"
                    name: "Appetizers"
                    time: "18:00"
                    description: "Light appetizers to start the evening"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Human-readable error message
          example: "Validation failed for one or more fields"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                description: Field that caused the error
              message:
                type: string
                description: Detailed error message for this field
            required:
              - field
              - message
          description: Detailed error information (optional)
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
        trace_id:
          type: string
          description: Unique identifier for tracing this error (optional)
      required:
        - error
        - message
        - timestamp
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          description: Current page number
          example: 1
        limit:
          type: integer
          minimum: 1
          description: Number of items per page
          example: 20
        total:
          type: integer
          minimum: 0
          description: Total number of items across all pages
          example: 45
        total_pages:
          type: integer
          minimum: 0
          description: Total number of pages
          example: 3
        has_next:
          type: boolean
          description: Whether there are more pages after this one
          example: true
        has_prev:
          type: boolean
          description: Whether there are pages before this one
          example: false
      required:
        - page
        - limit
        - total
        - total_pages
        - has_next
        - has_prev
    Address:
      type: object
      properties:
        address:
          type: string
          minLength: 1
          maxLength: 256
          description: Human-readable address
          example: "Marienplatz 1, 80331 München, Germany"
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
          description: Latitude coordinate (WGS84)
          example: 48.137154
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
          description: Longitude coordinate (WGS84)
          example: 11.575490
      required:
        - address
        - latitude
        - longitude
    CookAndRunMeta:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the cook and run project
        name:
          type: string
          maxLength: 64
          description: Name of the cook and run project
        created:
          type: string
          format: date-time
          description: When the project was created
        edited:
          type: string
          format: date-time
          description: When the project was last modified
        occur:
          type: string
          format: date-time
          description: When the event takes place (optional)
      required:
        - id
        - name
        - created
        - edited
        - occur
    CookAndRunCreateData:
      type: object
      required:
        - name
        - userId
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
          description: Name of the cook and run event
          example: "Summer Cook & Run 2024"
        userId:
          type: string
          maxLength: 64
          pattern: "^[a-zA-Z0-9_-]+"
          description: ID of the user creating the event
          example: "user123"
    CookAndRun:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the cook and run project
        name:
          type: string
          maxLength: 64
          description: Name of the cook and run project
        created:
          type: string
          format: date-time
          description: When the project was created
        edited:
          type: string
          format: date-time
          description: When the project was last modified
        occur:
          type: string
          format: date-time
          description: When the event takes place
        team_list:
          type: array
          items:
            $ref: "#/components/schemas/Team"
          description: List of registered teams
        course_list:
          type: array
          items:
            $ref: "#/components/schemas/Course"
          description: List of courses for the event
        start_point:
          allOf:
            - $ref: "#/components/schemas/Address"
          description: Starting location for the event (optional)
        end_point:
          allOf:
            - $ref: "#/components/schemas/Address"
          description: Ending location for the event (optional)
        share_team_config:
          allOf:
            - $ref: "#/components/schemas/ShareTeamConfig"
          description: Public team registration configuration (optional)
        plan:
          allOf:
            - $ref: "#/components/schemas/Plan"
          description: Generated event plan (optional)
      required:
        - id
        - name
        - created
        - edited
        - occur
        - team_list
        - course_list
...
    Team:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the team
        created_by_user:
          type: string
          maxLength: 64
          description: ID of the user who created this team
        name:
          type: string
          maxLength: 64
          description: Team name
        created:
          type: string
          format: date-time
          description: When the team was created
        edited:
          type: string
          format: date-time
          description: When the team was last modified
        address:
          $ref: "#/components/schemas/Address"
        mail:
          type: string
          format: email
          maxLength: 254
          description: Contact email for the team
        phone:
          type: string
          pattern: '^[\+]?[1-9][\d\s\-\(\)]{7,15}'
          maxLength: 20
          description: International phone number
        members:
          type: integer
          minimum: 1
          maximum: 20
          description: Number of team members
        diets:
          type: string
          maxLength: 256
          description: Special dietary requirements and allergies
        needs_check:
          type: boolean
          description: Whether team registration needs manual approval
        note_list:
          type: array
          items:
            $ref: "#/components/schemas/Note"
          description: Notes associated with this team
      required:
        - id
        - name
        - created
        - edited
        - address
        - needs_check
        - status
        - note_list
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token containing user authentication information.

        Token should include the following claims:
        - `sub`: User ID
        - `role`: User role (admin, organizer, participant)
        - `exp`: Token expiration time
        - `iss`: Token issuer (api.travelingcook.com)

        Format: `Authorization: Bearer <token>`

        Example: `Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`

# Security considerations:
# - Tokens should be refreshed every 24 hours
# - Use HTTPS in production
# - Implement proper token validation and blacklisting
# - Rate limiting is applied per authenticated user
...